+++ internal/config/config.go
@@ -33,6 +33,7 @@ type Defaults struct {
 	BaiduAppId               string `json:"baidu_app_id"`
 	BaiduSecretKey           string `json:"baidu_secret_key"`
 	CloseToTray              bool   `json:"close_to_tray"`
+	CompactMode              bool   `json:"compact_mode"`
 	CustomCssFile            string `json:"custom_css_file"`
 	DeeplAPIKey              string `json:"deepl_api_key"`
 	DeeplEndpoint            string `json:"deepl_endpoint"`
@@ -143,6 +144,8 @@ func GetString(key string) string {
 		return defaults.BaiduSecretKey
 	case "close_to_tray":
 		return strconv.FormatBool(defaults.CloseToTray)
+	case "compact_mode":
+		return strconv.FormatBool(defaults.CompactMode)
 	case "custom_css_file":
 		return defaults.CustomCssFile
 	case "deepl_api_key":