+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to MrRSS will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Supported compact mode for article list to reduce visual clutter. (#403)
+
 ## [1.3.15] - 2026-01-11
 
 ### Changed